gemini_term/requirements.txt
python-dotenv
rich
google-genai